Set pidfile location in puma config
By default, puma was putting the pidfile in the rails app `tmp`
directory. This was leading to the pid file existing when creating
a new container, thus making puma fail to start saying that there
was already a server running. Since we're running in containers we've
set the pid file to be in the `/tmp` directory that will start clean
with every container run.