Create Nginx frontend service for the rails application
This commit creates a custom Nginx image that can be used to run
an Nginx frontend service and deployment for the application. Additionally,
this commit fixes issues with the postgres deployment and storage
settings when working with minikube to prevent the database from being
lost when the postgres pod or minikube are stopped.